fastapi/learning/requirements-learning.txt
annotated-types
anyio
certifi
charset-normalizer
colorama
elastic-transport
elasticsearch
fastapi
filelock
fsspec
huggingface-hub
idna
Jinja2
uvicorn
loadenv
MarkupSafe
mpmath
networkx
numpy
packaging
pydantic
pydantic_core
python-dotenv
PyYAML
regex
requests
safetensors
scipy
setuptools
sniffio
starlette
sympy
tokenizers
torch
tqdm
transformers
typing_extensions
urllib3
